nivel casi ready

+++ working/prueba/polilinea.h
@@ -24,7 +24,7 @@ color_t polilinea_color(const polilinea_t *polilinea);
 //Setters
 bool polilinea_setear_punto(polilinea_t *polilinea, size_t pos, float x, float y);
 bool polilinea_setear_color(polilinea_t *polilinea, color_t color);
-void polilinea_trasladar(polilinea_t *polilinea, double traslacion[2]);
+void polilinea_trasladar(polilinea_t *polilinea, float traslacion[2]);
 void polilinea_rotar(polilinea_t *polilinea, double angulo);
 bool dibujar_polilinea(SDL_Renderer *renderer,polilinea_t *polilinea, float posicion[2], float escala);
